Factor out common fixture
